containers/video-server/requirements.txt
# Video Generation Server Dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Image/Video processing
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0

# AI/ML
torch>=2.1.0
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0

# HuggingFace
huggingface-hub>=0.20.0
